Move the common code for building an experiment to examples/support
